readora_backend/requirements.txt
django
gunicorn
whitenoise
django-cors-headers
dj-database-url
python-dotenv
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
daphne
cloudinary
requests
Pillow
uvicorn
